Drop compat code for `IOBuffer` from #501 and #504
